refactor: use grafana/data package Instead of using own implementation of the PanelData, use grafana/data package instead.
