+++ package.json
@@ -22,6 +22,7 @@
     "watch": "rollup -c rollup.config.dev.js -w"
   },
   "devDependencies": {
+    "@grafana/data": "^7.1.5",
     "@rollup/plugin-commonjs": "^17.1.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.2.0",
