+++ src/dev-site/create-series.ts
@@ -1,3 +1,5 @@
+import { DataFrame, FieldType } from "@grafana/data";
+
 function createSeries(name: string, value: number, refId = "A"): DataFrame {
   return {
     name: name,
