+++ src/dev-site/data.ts
@@ -1,3 +1,4 @@
+import { dateTime, LoadingState } from "@grafana/data";
 import createSeries from "./create-series";
 
 function randomValue({ min = 0, max = 1 }): number {
@@ -11,6 +12,14 @@ function updateData(): void {
       createSeries("random-series-1", randomValue({ min: 0, max: 100 })),
       createSeries("random-series-2", randomValue({ min: -100, max: 100 })),
     ],
+    timeRange: {
+      from: dateTime(0),
+      to: dateTime(0),
+      raw: {
+        from: dateTime(0),
+        to: dateTime(0),
+      },
+    },
   };
 }
