--- types/htmlgraphics-types/scoped-vars.d.ts
@@ -1,10 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-interface ScopedVar<T = any> {
-  text: any;
-  value: T;
-  [key: string]: any;
-}
-interface ScopedVars {
-  [key: string]: ScopedVar;
-}
