--- types/htmlgraphics-types/time-ranges.d.ts
@@ -1,93 +1,0 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
-interface DateTimeBuiltinFormat {
-  __momentBuiltinFormatBrand: any;
-}
-declare const ISO_8601: DateTimeBuiltinFormat;
-declare type DateTimeInput =
-  | Date
-  | string
-  | number
-  | Array<string | number>
-  | DateTime
-  | null;
-declare type FormatInput = string | DateTimeBuiltinFormat | undefined;
-declare type DurationInput = string | number | DateTimeDuration;
-declare type DurationUnit =
-  | "year"
-  | "years"
-  | "y"
-  | "month"
-  | "months"
-  | "M"
-  | "week"
-  | "weeks"
-  | "w"
-  | "day"
-  | "days"
-  | "d"
-  | "hour"
-  | "hours"
-  | "h"
-  | "minute"
-  | "minutes"
-  | "m"
-  | "second"
-  | "seconds"
-  | "s"
-  | "millisecond"
-  | "milliseconds"
-  | "ms"
-  | "quarter"
-  | "quarters"
-  | "Q";
-interface DateTimeLocale {
-  firstDayOfWeek: () => number;
-}
-interface DateTimeDuration {
-  asHours: () => number;
-  hours: () => number;
-  minutes: () => number;
-  seconds: () => number;
-  asSeconds: () => number;
-}
-interface DateTime extends Object {
-  add: (amount?: DateTimeInput, unit?: DurationUnit) => DateTime;
-  set: (unit: DurationUnit, amount: DateTimeInput) => void;
-  diff: (
-    amount: DateTimeInput,
-    unit?: DurationUnit,
-    truncate?: boolean
-  ) => number;
-  endOf: (unitOfTime: DurationUnit) => DateTime;
-  format: (formatInput?: FormatInput) => string;
-  fromNow: (withoutSuffix?: boolean) => string;
-  from: (formaInput: DateTimeInput) => string;
-  isSame: (input?: DateTimeInput, granularity?: DurationUnit) => boolean;
-  isBefore: (input?: DateTimeInput) => boolean;
-  isValid: () => boolean;
-  local: () => DateTime;
-  locale: (locale: string) => DateTime;
-  startOf: (unitOfTime: DurationUnit) => DateTime;
-  subtract: (amount?: DateTimeInput, unit?: DurationUnit) => DateTime;
-  toDate: () => Date;
-  toISOString: () => string;
-  isoWeekday: (day?: number | string) => number | string;
-  valueOf: () => number;
-  unix: () => number;
-  utc: () => DateTime;
-  utcOffset: () => number;
-  hour?: () => number;
-  minute?: () => number;
-}
-
-interface RawTimeRange {
-  from: DateTime | string;
-  to: DateTime | string;
-}
-
-interface TimeRange {
-  from: DateTime;
-  to: DateTime;
-  raw: RawTimeRange;
-}
