--- types/window.d.ts
@@ -1,7 +1,0 @@
-interface Window {
-  htmlNode: typeof htmlNode;
-  customProperties: typeof customProperties;
-  data: typeof data;
-  options: typeof options;
-  theme: typeof theme;
-}
